Files to be used for creating our game
These are the files that we will use to create our game, each file is named based on the code that will be inside it. When you upload your version of that file, put the date of when you made/edited it, so that we can make sure we have all the up to date files.